Small fixes; add engagement
